split NivoCharts.js into multiple files
